Move to 64-bit
Everything should be properly converted, and definitions use a set of
macros now to manage the links and labels automatically.

This macro system was inspired by Itsy Forth, but modified for
subroutine threading and 64 bits. Additionally, there's only one word to
create definitions, since colon and code definitions are the same under
STC, and the macros for variables and constants simply integrate their
corresponding code.